apps/consumer/requirements.txt
astroid==2.3.3
attrs==19.3.0
daemonize==2.5.0
docutils==0.15.2
importlib-metadata==1.3.0
isort==4.3.21
kafka-python==1.4.7
lazy-object-proxy==1.4.3
lockfile==0.12.2
mccabe==0.6.1
more-itertools==8.0.2
packaging==20.0
pluggy==0.13.1
port-for==0.4
py==1.8.1
pylint==2.4.4
pyparsing==2.4.6
pytest==5.3.2
pytest-kafka==0.4.0
six==1.13.0
statsd==3.3.0
typed-ast==1.4.0
wcwidth==0.1.8
wrapt==1.11.2
zipp==0.6.0
